Amendment to Standardize the Bylaws
WHEREAS, the text of the Bylaws is not readily apparent in its current form; and
WHEREAS, having a standardized current version of the Bylaws is crucial to the well functioning of the Union; now, therefore, be it
RESOLVED, that the Bylaws be amended to read as proposed herein.

Sponsored by Abhishek Kulkarni, Senator for the Class of 2018.