Amendment to Open Bylaws Committee to non-Senators
WHEREAS, the Student Union Senate is responsible for maintaining open lines of communication with the student body,
WHEREAS, the bylaws committee is responsible for enhancing the efficiency of the Senate,
WHEREAS, the bylaws committee lacks input from non-Senate members,
THEREFORE, be it resolved that Article VI, Section 6 of the Student Union Bylaws be amended as follows. 

Sponsored by Jake Rong, Senator for Village and 567, and Alex Chang, Senator for the Class of 2022.