Amendment to Add A New Section to Article V
WHEREAS, the Student Union Senate is responsible for voting on proposals,
WHEREAS, the rules committee is responsible for enhancing the efficiency of the senate,
WHEREAS, repeated vote changes inhibit the timely succession of agenda items,
THEREFORE, be it resolved that Article V of the Student Union Bylaws be amended as follows. 

Sponsored by Jake Rong, Senator for Village and 567, and Leigh Salomon, Senator for Ziv and Ridgewood.