Amendment Requiring Judiciary Attendance at Rules Committee
WHEREAS, there is a compelling need to improve transparency among branches,
WHEREAS, the rules committee is responsible for maintaining the Bylaws,
WHEREAS, the Judiciary is responsible for interpreting the Bylaws,
WHEREAS, the Judiciary will be more aware of the Bylaws by sitting in on rules committee Meetings,
WHEREAS, the Judiciary is ultimately responsible for repealing Bylaws that are unconstitutional,
THEREFORE, be it resolved that Article VI, Section 6 of the Student Union Bylaws be amended as follows. 

Sponsored by Jake Rong, Senator for Village and 567, and Leigh Salomon, Senator for Ziv and Ridgewood.