Amendment to Eliminate the Club Website Requirement
WHEREAS, the Student Union Senate is responsible for maintaining the Bylaws pertaining to clubs,
WHEREAS, the club websites have fallen out of use,
WHEREAS, the rules committee is responsible for enhancing the clarity of the Bylaws,
THEREFORE, be it resolved that Article VIII, Section 7 of the Student Union Bylaws be amended as follows. 

Sponsored by Jake Rong, Senator for Village and 567, and Leigh Salomon, Senator for Ziv and Ridgewood.