Amendment Expanding Committee Chair Attendance Record to Branch Heads
WHEREAS, Senate Committee Chairs are required to take attendance at meetings,
WHEREAS, Branch Heads should be held to the same standard as Committee Chairs,
WHEREAS, the Bylaws fail to mandate records of attendance for non-Senators,
THEREFORE, be it resolved that Article II, Section 5 be added to the Student Union Bylaws, and that Article VI, Section 3 be removed from the Student Union Bylaws. 

Sponsored by Jake Rong, Senator for Village and 567, and Leigh Salomon, Senator for Ziv and Ridgewood.