Amendment Clarifying A-Board Policy
WHEREAS, the Brandeis Allocations Board is responsible for approving the financial requests of Chartered Student Clubs, 
WHEREAS, the rules committee is responsible for promoting clarity within the Student Union,
WHEREAS, ambiguity about the Allocations Board’s responsibilities leads to  discord within the Brandeis Student Union and the community it represents,
THEREFORE, be it resolved that Article III of the Student Union Bylaws be amended as follows. 

Sponsored by Jake Rong, Senator for Village and 567, Leigh Salomon, Senator for Ziv and Ridgewood, and Will Hodgkinson, Class of 2021.