Amendment Giving the Senate the Power to Subpoena Club Leaders
WHEREAS, the rules committee is responsible for maintaining the Bylaws pertaining to clubs,
WHEREAS the Student Union Senate is responsible for ensuring that clubs follow the Bylaws,
WHEREAS, club leaders are responsible for their clubs, 
THEREFORE, be it resolved that Article VIII, Section 8 of the Student Union Bylaws be amended as follows. 

Sponsored by Jake Rong, Senator for Village and 567, and Leigh Salomon, Senator for Ziv and Ridgewood.